move msToTime() to \libs\utils\time
