+++ app/lib/utils/time.ts
@@ -41,3 +41,20 @@ export const timeDifference = (current: number, previous: number): string => {
     const time = Math.round(elapsed / msPerYear);
     return `${time} year${addPlural(time)} ago`;
 };
+
+export const msToTime = (duration: number) => {
+    const seconds = Math.floor((duration / 1000) % 60);
+    const minutes = Math.floor((duration / (1000 * 60)) % 60);
+    const hours = Math.floor((duration / (1000 * 60 * 60)) % 24);
+
+    if (hours) {
+        return `${hours} hours and ${minutes} minutes`;
+    }
+    if (minutes) {
+        return `${minutes} minutes and ${seconds} seconds`;
+    }
+    if (seconds) {
+        return `${seconds} seconds`;
+    }
+    return ('');
+};
